+++ src/Phan/Analysis/ConditionVisitor.php
@@ -306,73 +306,117 @@ class ConditionVisitor extends KindVisitorImplementation implements ConditionVis
      */
     public function visitIsset(Node $node) : Context
     {
-        $context = $this->context;
         $var_node = $node->children['var'];
         if (!($var_node instanceof Node)) {
-            return $context;
+            return $this->context;
         }
-        if (($var_node->kind ?? null) !== ast\AST_VAR) {
+        if ($var_node->kind !== ast\AST_VAR) {
             return $this->checkComplexIsset($var_node);
         }
 
         $var_name = $var_node->children['name'];
         if (!\is_string($var_name)) {
             $this->checkVariablesDefinedInIsset($var_node);
-            return $context;
+            return $this->context;
         }
-        if (!$context->getScope()->hasVariableWithName($var_name)) {
+        return $this->withSetVariable($var_name, $var_node);
+    }
+
+    /**
+     * From isset($var), infer that $var is non-null
+     * From isset($obj->prop['field']), infer that $obj is non-null
+     * Also infer that $obj is an object (don't do that for $obj['field']->prop)
+     */
+    private function withSetVariable(string $var_name, Node $var_node) : Context
+    {
+        $context = $this->context;
+        $is_object = $var_node->kind === ast\AST_PROP;
+
+        $scope = $context->getScope();
+        if (!$scope->hasVariableWithName($var_name)) {
             // Support analyzing cases such as `if (isset($x)) { use($x); }`, or `assert(isset($x))`
-            $context->setScope($context->getScope()->withVariable(new Variable(
+            return $context->withScopeVariable(new Variable(
                 $context->withLineNumberStart($var_node->lineno ?? 0),
                 $var_name,
-                UnionType::empty(),
-                $var_node->flags
-            )));
-            return $context;
+                $is_object ? ObjectType::instance(false)->asUnionType() : UnionType::empty(),
+                0
+            ));
+        }
+        if ($is_object) {
+            $variable = clone($context->getScope()->getVariableByName($var_name));
+            $this->analyzeIsObjectAssertion($variable);
+            return $context->withScopeVariable($variable);
         }
         return $this->removeNullFromVariable($var_node, $context, true);
     }
 
-    private function checkComplexIsset(Node $var_node) : Context
+    /**
+     * @param Node $node a node that is NOT of type ast\AST_VAR
+     */
+    private function checkComplexIsset(Node $node) : Context
     {
-        // TODO: isset($obj->prop['offset']) should imply $obj is not null (removeNullFromVariable)
+        // Loop to support getting the var name in is_array($x['field'][0])
+        $has_prop_access = false;
         $context = $this->context;
-        if ($var_node->kind === ast\AST_DIM) {
-            $expr_node = $var_node;
-            do {
-                $parent_node = $expr_node;
-                $expr_node = $expr_node->children['expr'];
-                if (!($expr_node instanceof Node)) {
-                    return $context;
-                }
-            } while ($expr_node->kind === ast\AST_DIM);
-
-            if ($expr_node->kind === ast\AST_VAR) {
-                $var_name = $expr_node->children['name'];
-                if (!\is_string($var_name)) {
-                    return $context;
-                }
-                if (!$context->getScope()->hasVariableWithName($var_name)) {
-                    // Support analyzing cases such as `if (isset($x['key'])) { use($x); }`, or `assert(isset($x['key']))`
-                    $context->setScope($context->getScope()->withVariable(new Variable(
-                        $context->withLineNumberStart($expr_node->lineno ?? 0),
-                        $var_name,
-                        ArrayType::instance(false)->asUnionType(),
-                        $expr_node->flags
-                    )));
-                    return $context;
-                }
-                $context = $this->removeNullFromVariable($expr_node, $context, true);
-
-                $variable = $context->getScope()->getVariableByName($var_name);
-                $var_node_union_type = $variable->getUnionType();
-
-                if ($var_node_union_type->hasTopLevelArrayShapeTypeInstances()) {
-                    $context = $this->withSetArrayShapeTypes($variable, $parent_node->children['dim'], $context, true);
-                }
-                $this->context = $context;
+        $var_node = $node;
+        $parent_node = $node;
+        while (true) {
+            if (!($var_node instanceof Node)) {
+                return $context;
             }
+            $kind = $var_node->kind;
+            if ($kind === ast\AST_VAR) {
+                break;
+            }
+            $parent_node = $var_node;
+            if ($kind === ast\AST_DIM) {
+                $var_node = $var_node->children['expr'];
+                if (!$var_node instanceof Node) {
+                    return $context;
+                }
+                continue;
+            } elseif ($kind == ast\AST_PROP) {
+                $has_prop_access = true;
+                $var_node = $var_node->children['expr'];
+                if (!$var_node instanceof Node) {
+                    return $context;
+                }
+                continue;
+            }
+
+            // TODO: Handle more than one level of nesting
+            return $context;
         }
+        $var_name = $var_node->children['name'];
+        if (!is_string($var_name)) {
+            return $context;
+        }
+        if ($has_prop_access) {
+            // For `$x->prop['field'][0]`, $parent_node would be `$x->prop`.
+            // And for that expression, phan would infer that $var_name was non-null AND an object.
+            return $this->withSetVariable($var_name, $parent_node);
+        }
+
+        // This is $x['field'] or $x[$i][something]
+
+        if (!$context->getScope()->hasVariableWithName($var_name)) {
+            // Support analyzing cases such as `if (isset($x['key'])) { use($x); }`, or `assert(isset($x['key']))`
+            return $context->withScopeVariable(new Variable(
+                $context->withLineNumberStart($node->lineno ?? 0),
+                $var_name,
+                ArrayType::instance(false)->asUnionType(),
+                0
+            ));
+        }
+        $context = $this->removeNullFromVariable($var_node, $context, true);
+
+        $variable = $context->getScope()->getVariableByName($var_name);
+        $var_node_union_type = $variable->getUnionType();
+
+        if ($var_node_union_type->hasTopLevelArrayShapeTypeInstances()) {
+            $context = $this->withSetArrayShapeTypes($variable, $parent_node->children['dim'], $context, true);
+        }
+        $this->context = $context;
         return $context;
     }